backend/tools/search_ai.requirements.txt
faster-whisper>=1.0.3
Pillow>=10.4.0
torch>=2.4.0
transformers>=4.46.0
